requirements.txt
pydantic==1.8.2
tinydb==4.5.1
python-dateutil